refactor(listings): retire Batch 1 listing proxy files and align SSOT imports
