+++ core/src/domains/listings/application/policies/ListingSubmissionPolicy.ts
@@ -1,5 +1,5 @@
 import type { ClientSession } from 'mongoose';
-import { AdSlotService, type AdPostingSlotSource } from '../../../../services/AdSlotService';
+import { AdSlotService, type AdPostingSlotSource } from '../../../boosts/application/services/AdSlotService';
 import { LISTING_TYPE, type ListingTypeValue } from '@esparex/contracts';
 import { checkPostLimit } from '../../../../services/PlanService';
 import type { ListingRepositoryPort } from '../../../../domains/listings';
