--- core/src/services/AdSlotService.ts
@@ -1,1 +1,0 @@
-export * from "../domains/boosts/application/services/AdSlotService";
